Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pymysql
openai
whisper
ffmpeg-python
python-dotenv
pydub
vosk
google-generativeai
PyPDF2 
python-docx
